Initial implementaion of board and stack
board is the representation of a chess board. Need to change the
memory type from flip-flops to BRAM.
stack remembers the last 3 moves played and exposes the head of the
stack. It also supports push and pop operations.

Used verilog unpacked structures to implement them, going to OH tomorrow
to fully understand their meaning

Also included is sim.sh which runs testbenches in an easier format